+++ components/elm/src/biogeochem/ChemStateType.F90
@@ -49,7 +49,7 @@ module ChemStateType
     type(bounds_type), intent(in)    :: bounds  
     !
     ! !LOCAL VARIABLES:
-    integer :: begp, endp
+    integer*8 :: begp, endp
     integer :: begc, endc
     integer :: begg, endg
     integer :: lbj,  ubj
