+++ components/elm/src/biogeophys/SoilHydrologyType.F90
@@ -110,7 +110,7 @@ contains
     type(bounds_type), intent(in) :: bounds
     !
     ! !LOCAL VARIABLES:
-    integer :: begp, endp
+    integer*8 :: begp, endp
     integer :: begc, endc
     integer :: begg, endg
     !------------------------------------------------------------------------
@@ -171,7 +171,7 @@ contains
     type(bounds_type), intent(in) :: bounds
     !
     ! !LOCAL VARIABLES:
-    integer           :: begp, endp
+    integer*8           :: begp, endp
     integer           :: begc, endc
     integer           :: begg, endg
     !------------------------------------------------------------------------
