+++ components/elm/src/biogeophys/SurfaceAlbedoType.F90
@@ -240,7 +240,7 @@ contains
     type(bounds_type), intent(in) :: bounds
     !
     ! !LOCAL VARIABLES:
-    integer :: begp, endp
+    integer*8 :: begp, endp
     integer :: begc, endc
     !---------------------------------------------------------------------
 
@@ -308,7 +308,7 @@ contains
     type(bounds_type), intent(in) :: bounds
     !
     ! !LOCAL VARIABLES:
-    integer :: begp, endp
+    integer*8 :: begp, endp
     integer :: begc, endc
     !---------------------------------------------------------------------
 
@@ -352,7 +352,7 @@ contains
     type(bounds_type), intent(in) :: bounds
     !
     ! !LOCAL VARIABLES:
-    integer :: begc, endc
+    integer*8 :: begc, endc
     integer :: begp, endp
     !-----------------------------------------------------------------------
 
@@ -415,7 +415,7 @@ contains
     ! !LOCAL VARIABLES:
     logical :: readvar      ! determine if variable is on initial file
     integer :: iv
-    integer :: begp, endp
+    integer*8 :: begp, endp
     integer :: begc, endc
     !---------------------------------------------------------------------
